Moved Vertex struct into separate file
